Adding ninjWAS module and 2017 Course Data
Added ninjWAS module for scraping Web Advisor. Takes cares of most exceptions so it will **almost** never crash. It provides a lot of logging directly to stdout or to a log file if specified that should tell you where the error is occurring, but most issues can be resolved by just rerunning the program.